requirements.txt
Django==1.11.4
django-appconf==1.0.2
django-imagekit==4.0.1
mysqlclient==1.3.10
pilkit==2.0
pytz==2017.2
six==1.10.0
South==1.0.2
